Optimize the runOpenApi plugin
allow passing in an already dereferenced OpenAPI Document for plugins
that have large openapi specs like github. This takes the runtime for
Github from over 10s to ~1s on an M1 Max.

Adjust the github plugin to use this new capability by dereferencing on
plugin initialization, passing the derefed document to the runOpenApi
function.