Add a filter to the New Data Source modal
add a filter to make it easier to find the data source you're looking
for in the New Data Source modal found in both the lower pane when
creating an app, and the Data Source page when in Home.

The filtering ignores case and diacritics, but is not a fuzzy filter as
I thought that would be too confusing for a simple search like this.

The localeContains function comes from StackOverflow user Jan Turoň, https://stackoverflow.com/a/69623589/1394698

The function sacrifices complete compatibility with the localeCompare
function for a simplified diacritics check which should suffice for the
majority of lowcoder use cases. I initially tried to get the npm library
locale-includes working, but Vite really doesn't like it, and it failed
to compile no matter what I tried.