Fix two small bugs with the RestAPI Datasource
First bug:

Spring automatically transforms binary content to base64, therefore the
logic in the RestApiExecutor was causing the data to be encoded twice,
causing much confusion and frustration for several weeks actually.

Second bug:

the fileUtils#saveDataAsFile function was not actually setting the type
correctly when attempting to convert the base64 data to a binary blob,
thus the files would _always_ download as a 'dms' file and it would
never actually be openable. This also caused a lot of confusion. Both of
these bugs together made it impossible to tell why stuff was broken.

Combined this means that anytime you were trying to download binary data
you would never actually be able to open the file, as it would be only
base64 data inside of it.