Allow pinning the code editor popup
The code editor popup now has a button to allow 'pinning' the editor to
the screen so that it does not disappear when you click outside of it.
This currently only works if the right panel stays open, so that means
you cannot click inside of the app window unless it's on your component.
Anything else will dismiss the right panel and the popup will disappear,
even if pinned.