Get tour step refs to persist past reload
I don't know exactly what I was doing wrong, so I copied the dropdown
control and modified it to accept the editor state. this could most
likely be done cleaner by allowing for the existing dropdown control to
accept a function with generic arguments, but that can be cleaned up
later.